Add sockaddr_t to fsal_getclientcontext
This patch adds a sockaddr_t* parameter to the FSAL_GetClientContext
function, and also the fsal_getclientcontext function pointer.
A sockaddr_t is also added to the user_credentials structure.
This will enable fsals to copy the ip into client context when
fsal_getclientcontext is called.

Signed-off-by: Allison Henderson <achender@linux.vnet.ibm.com>
------------------
v1 -> v2
  Instead of passing the entire svc_req,
  we now only pass the sockaddr_t

v2 -> v3
  Merged patches for FSAL_GetClientContext and fsal_getclientcontext