Remove GPFS Dependency
This patch removes the GPFS package dependency from the
PTFSAL.  The PT fsal was originally modeled after the GPFS fsal,
and some references to the gpfs libraries were inadvertently left
in some of the PT fsal routines.  The references were later found
when installing ganesha for PT on a clean box, and dependency
errors were encountered.

This patch removes the file FSAL_PT/fsal_lock.c.  This file only had
one function in it, PTFSAL_lock_op, and it needed to be removed.
The existing routine was filling out some gpfs structures,
but they were not being used for any functional purpose as they were in
the GPFSFSAL_lock_op routine.  Sense PT does not support locking right
now, the function and the file are removed, and the fsal_lock_op function
pointer is set to NULL.  It can be filled in properly when
PT is made to support locking at a later time.

This patch also removes the "attr_valid" and "attr_changed" variables
from the PTFSAL_setattrs.  These variables were initially used to
store GPFS specific attributes, and then used as parameters to another
GPFS function that is not present in the PT fsal.  The existing variables
were still referencing GPFS attributes, but sense they are not used
for anything, they are now simply removed.

GPFS header includes and library linkers are also removed.

Signed-off-by: Allison Henderson <achender@linux.vnet.ibm.com>
-----------
v1 -> v2
Removed some extra white space

v2->v3
Removed the entire FSAL_PT/fsal_lock.c sense it only had the
on function in it