Fix hard log levels
The current PT FSAL does not use the standard ganesha logging
routines.  It's log level is currently hard coded, and cannot
change with out recompiling.  The PT FSAL logging scheme also
does not have the same number of levels as the ganesha scheme,
so we need some sort for translation to allow it to use the
configurations passed on the command line.

This patch adds a new routine "ptfsal_convert_log_level"
that translates a ganesha log level to an fsi log level.
The old hard coded log level has been moved to the init
script.  All log levels appearing in the configs have also
been moved to the init script sense that is the currently
accepted method of setting log levels.

Signed-off-by: Allison Henderson <achender@linux.vnet.ibm.com>